pyproject.toml
[tool.poetry]
name = "iih-data-scraper"
version = "0.1.0"
description = ""
authors = ["abdalkreemorabi <abdalkreemorabi11@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
bs4 = "^0.0.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
